pyproject.toml
[project]
name = "glimpse"
version = "0.1.0"
description = "macOS screenshot capture, OCR, and intelligence layer with REST API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pillow>=11.0",
    "numpy>=2.0",
    "pyobjc-core>=11.0",
    "pyobjc-framework-Quartz>=11.0",
    "pyobjc-framework-Vision>=11.0",
    "pyobjc-framework-ApplicationServices>=11.0",
    "pyobjc-framework-Cocoa>=11.0",
    "pydantic>=2.0",
    "aiosqlite>=0.20",
]

[project.scripts]
glimpse = "glimpse.main:main"

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
